Add tool used for bisecting troublesome tests (eg. looping, non-deterministic)
The tool returns 0 if the configured test can pass.

There is a timeout that can be set to fail looping tests.

For non-deterministic use, a test can be considered passed after a number of
iterations.

The script requires tuning for each bisection.

Further work in future could improve it to query the user for parameters on
first start, save these to a local file and read the parameters from that
file if present.
This would avoid the need to change this script during testing.